+++ src/main/java/net/wanho/service/base/BaseService.java
@@ -2,15 +2,15 @@ package net.wanho.service.base;
 
 public interface BaseService<T> {
 
-    Long deleteByPrimaryKey(Long tid);
+    int deleteByPrimaryKey(Long tid);
 
-    Long insert(T record);
+    int insert(T record);
 
-    Long insertSelective(T record);
+    int insertSelective(T record);
 
     T selectByPrimaryKey(Long tid);
 
-    Long updateByPrimaryKeySelective(T record);
+    int updateByPrimaryKeySelective(T record);
 
-    Long updateByPrimaryKey(T record);
+    int updateByPrimaryKey(T record);
 }
